+++ src/extension.ts
@@ -14,16 +14,16 @@
  *   limitations under the License.
  */
 
-import * as vscode from "vscode";
-import * as Constants from "./Constants";
-import { ConfigurationManager } from "./ConfigurationManager";
 import { IAnnotatedtext } from "annotatedtext";
-import { Linter } from "./Linter";
+import * as vscode from "vscode";
+import { ConfigurationManager } from "./ConfigurationManager";
+import * as Constants from "./Constants";
 import { FormattingProviderDashes } from "./FormattingProviderDashes";
-import { OnTypeFormattingDispatcher } from "./OnTypeFormattingDispatcher";
 import { FormattingProviderEllipses } from "./FormattingProviderEllipses";
 import { FormattingProviderQuotes } from "./FormattingProviderQuotes";
 import { ILanguageToolMatch } from "./Interfaces";
+import { Linter } from "./Linter";
+import { OnTypeFormattingDispatcher } from "./OnTypeFormattingDispatcher";
 
 // Wonder Twin Powers, Activate!
 export function activate(context: vscode.ExtensionContext): void {
@@ -42,7 +42,7 @@ export function activate(context: vscode.ExtensionContext): void {
   context.subscriptions.push(Constants.EXTENSION_OUTPUT_CHANNEL);
   Constants.EXTENSION_OUTPUT_CHANNEL.appendLine("LanguageTool Linter Activated!");
 
-  // Register onDidChangeconfiguration event
+  // Register onDidChangeConfiguration event
   context.subscriptions.push(
     vscode.workspace.onDidChangeConfiguration((event) => {
       if (event.affectsConfiguration("languageToolLinter")) {
@@ -54,39 +54,24 @@ export function activate(context: vscode.ExtensionContext): void {
   // Register onDidOpenTextDocument event - request lint
   context.subscriptions.push(
     vscode.workspace.onDidOpenTextDocument((document) => {
-      if (configMan.isLintOnOpen()) {
-        linter.requestLint(document);
-      }
+      linter.documentChanged(document, configMan.isLintOnOpen());
     }),
   );
 
   // Register onDidChangeTextDocument event - request lint with default timeout
   context.subscriptions.push(
     vscode.workspace.onDidChangeTextDocument((event) => {
-      if (configMan.isLintOnChange()) {
-        if (configMan.isHideDiagnosticsOnChange()) {
-          linter.clearDiagnostics(event.document.uri);
-        }
-        linter.requestLint(event.document);
-      }
+      linter.documentChanged(event.document, configMan.isLintOnChange());
     }),
   );
 
   // Register onDidChangeActiveTextEditor event - request lint
   context.subscriptions.push(
     vscode.window.onDidChangeActiveTextEditor((editor) => {
-      if (editor !== undefined && configMan.isLintOnChange()) {
-        linter.requestLint(editor.document);
-      }
-    }),
-  );
-
-  // Register onWillSaveTextDocument event - smart format if enabled
-  context.subscriptions.push(
-    vscode.workspace.onWillSaveTextDocument((_event) => {
-      if (configMan.isSmartFormatOnSave()) {
-        vscode.commands.executeCommand(
-          "languagetoolLinter.smartFormatDocument",
+      if (editor !== undefined) {
+        linter.editorChanged(
+          vscode.window.activeTextEditor,
+          configMan.isLintOnChange(),
         );
       }
     }),
@@ -95,8 +80,15 @@ export function activate(context: vscode.ExtensionContext): void {
   // Register onDidSaveTextDocument event - request immediate lint
   context.subscriptions.push(
     vscode.workspace.onDidSaveTextDocument((document) => {
-      if (configMan.isLintOnSave()) {
-        linter.requestLint(document);
+      linter.documentChanged(document, configMan.isLintOnSave());
+    }),
+  );
+
+  // Register onWillSaveTextDocument event - smart format if enabled
+  context.subscriptions.push(
+    vscode.workspace.onWillSaveTextDocument((_event) => {
+      if (configMan.isSmartFormatOnSave()) {
+        vscode.commands.executeCommand(Constants.COMMAND_SMART_FORMAT);
       }
     }),
   );
@@ -137,9 +129,38 @@ export function activate(context: vscode.ExtensionContext): void {
     }),
   );
 
+  // Register "Toggle Linting" Command
+  const toggleLinting = vscode.commands.registerCommand(
+    Constants.COMMAND_TOGGLE_LINTING,
+    () => {
+      linter.toggleSuspendLinting();
+    },
+  );
+  context.subscriptions.push(toggleLinting);
+
+  // Register "Disable Rule" TextEditorCommand
+  const disableRule = vscode.commands.registerTextEditorCommand(
+    Constants.COMMAND_DISABLE_RULE,
+    (editor, _edit, ...args) => {
+      configMan.disableRule(args.shift(), args.shift());
+      linter.requestLint(editor.document, 0);
+    },
+  );
+  context.subscriptions.push(disableRule);
+
+  // Register "Disable Category" TextEditorCommand
+  const disableCategory = vscode.commands.registerTextEditorCommand(
+    Constants.COMMAND_DISABLE_CATEGORY,
+    (editor, _edit, ...args) => {
+      configMan.disableCategory(args.shift(), args.shift());
+      linter.requestLint(editor.document, 0);
+    },
+  );
+  context.subscriptions.push(disableCategory);
+
   // Register "Ignore Word Globally" TextEditorCommand
   const ignoreWordGlobally = vscode.commands.registerTextEditorCommand(
-    "languagetoolLinter.ignoreWordGlobally",
+    Constants.COMMAND_IGNORE_USR_WORD,
     (editor, _edit, ...args) => {
       configMan.ignoreWordGlobally(args.shift());
       linter.requestLint(editor.document, 0);
@@ -149,7 +170,7 @@ export function activate(context: vscode.ExtensionContext): void {
 
   // Register "Ignore Word in Workspace" TextEditorCommand
   const ignoreWordInWorkspace = vscode.commands.registerTextEditorCommand(
-    "languagetoolLinter.ignoreWordInWorkspace",
+    Constants.COMMAND_IGNORE_WS_WORD,
     (editor, _edit, ...args) => {
       configMan.ignoreWordInWorkspace(args.shift());
       linter.requestLint(editor.document, 0);
@@ -187,7 +208,7 @@ export function activate(context: vscode.ExtensionContext): void {
 
   // Register "Remove Globally Ignored Word" TextEditorCommand
   const removeGloballyIgnoredWord = vscode.commands.registerTextEditorCommand(
-    "languagetoolLinter.removeGloballyIgnoredWord",
+    Constants.COMMAND_REMOVE_USR_IGNORED_WORD,
     (editor, _edit, ...args) => {
       configMan.removeGloballyIgnoredWord(args.shift());
       linter.requestLint(editor.document, 0);
@@ -197,7 +218,7 @@ export function activate(context: vscode.ExtensionContext): void {
 
   // Register "Remove Workspace Ignored Word" TextEditorCommand
   const removeWorkspaceIgnoredWord = vscode.commands.registerTextEditorCommand(
-    "languagetoolLinter.removeWorkspaceIgnoredWord",
+    Constants.COMMAND_REMOVE_WS_IGNORED_WORD,
     (editor, _edit, ...args) => {
       configMan.removeWorkspaceIgnoredWord(args.shift());
       linter.requestLint(editor.document, 0);
@@ -207,7 +228,7 @@ export function activate(context: vscode.ExtensionContext): void {
 
   // Register "Check Current Document" TextEditorCommand
   const checkDocument = vscode.commands.registerTextEditorCommand(
-    "languagetoolLinter.checkDocument",
+    Constants.COMMAND_CHECK_DOCUMENT,
     (editor: vscode.TextEditor, _edit: vscode.TextEditorEdit) => {
       linter.requestLint(editor.document, 0);
     },
@@ -216,7 +237,7 @@ export function activate(context: vscode.ExtensionContext): void {
 
   // Register "Check as Plain Text Document" TextEditorCommand
   const checkDocumentAsPlainText = vscode.commands.registerTextEditorCommand(
-    "languagetoolLinter.checkDocumentAsPlainText",
+    Constants.COMMAND_CHECK_DOCUMENT_AS_PLAINTEXT,
     (editor: vscode.TextEditor, _edit: vscode.TextEditorEdit) => {
       linter.requestLintAsPlainText(editor.document, 0);
     },
@@ -240,7 +261,7 @@ export function activate(context: vscode.ExtensionContext): void {
 
   // Register "Clear LanguageTool Diagnostics" TextEditorCommand
   const clearDocumentDiagnostics = vscode.commands.registerTextEditorCommand(
-    "languagetoolLinter.clearDocumentDiagnostics",
+    Constants.COMMAND_CLEAR_DIAGNOSTICS,
     (editor: vscode.TextEditor, _edit: vscode.TextEditorEdit) => {
       linter.clearDiagnostics(editor.document.uri);
     },
@@ -249,7 +270,7 @@ export function activate(context: vscode.ExtensionContext): void {
 
   // Register "Smart Format Document" TextEditorCommand
   const smartFormatCommand = vscode.commands.registerTextEditorCommand(
-    "languagetoolLinter.smartFormatDocument",
+    Constants.COMMAND_SMART_FORMAT,
     (editor: vscode.TextEditor, edit: vscode.TextEditorEdit) => {
       if (configMan.isSupportedDocument(editor.document)) {
         // Revert to regex here for cleaner code.
@@ -283,7 +304,6 @@ export function activate(context: vscode.ExtensionContext): void {
   }
 }
 
-// eslint-disable-next-line @typescript-eslint/no-empty-function
 export function deactivate(): void {}
 
 /**